Rearrange attribution protocol a little
This is still a bit janky in that the Step here is a collection of
random stuff.  But this just merges the malicious implementation down
into the top level.
