+++ src/protocol/attribution/attribution.rs
@@ -1,0 +1,32 @@
+use super::{
+    accumulate_credit::accumulate_credit,
+    aggregate_credit::aggregate_credit,
+    apply_attribution_window::apply_attribution_window,
+    compute_helper_bits_gf2, compute_stop_bits,
+    credit_capping::credit_capping,
+    input::{
+        MCAggregateCreditOutputRow, MCApplyAttributionWindowInputRow,
+        MCCappedCreditsWithAggregationBit,
+    },
+    mod_conv_helper_bits,
+};
+use crate::{
+    error::Error,
+    ff::{GaloisField, Gf2, PrimeField, Serializable},
+    helpers::query::IpaQueryConfig,
+    protocol::{
+        boolean::RandomBits,
+        context::{Context, UpgradableContext, UpgradedContext, Validator},
+        ipa::IPAModulusConvertedInputRow,
+        sort::generate_permutation::ShuffledPermutationWrapper,
+        BasicProtocols, Substep,
+    },
+    secret_sharing::{
+        replicated::{
+            malicious::{DowngradeMalicious, ExtendableField},
+            semi_honest::AdditiveShare as SemiHonestAdditiveShare,
+        },
+        Linear as LinearSecretSharing,
+    },
+};
+use std::iter::{once, zip};
